middleware: create a logger on each request
The AddContext() method added in #26 made it possible for HTTP middleware to
add additional keys and values to the logs generated by the kayvee
middleware logger. This is useful to log request-specific error messages or
IDs.

Since kayvee middleware uses a single shared logger object across all
requests, there are two problems with using AddContext:

- `AddContext` builds up context that never goes away. E.g., if a
  request fails and the handler puts the error reason in the kayvee
  logger context, it will be logged on subsequent requests.

- `AddContext` writes to a map, which will panic the process when done
  concurrently from multiple requests.

This commit changes kayvee middleware to create a new Logger object on
each request, which makes it possible for http handlers to safely call `AddContext`
on this logger.